# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InterRPA Local API",
    "description": "Sistema offline de gestión de envíos para oficinas locales. Escaneo persistente, cola de trabajos automática y sincronización con APIs públicas.",
    "version": "0.1.0",
    "contact": {
      "name": "InterRPA Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/zones": {
      "get": {
        "summary": "Listar zonas",
        "description": "Obtiene todas las zonas de entrega, con filtro opcional por estado activo",
        "operationId": "getZones",
        "tags": ["Zones"],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "schema": { "type": "integer", "enum": [0, 1] },
            "description": "Filtrar por estado activo (1) o inactivo (0)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de zonas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "zones": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" },
                          "active": { "type": "integer", "enum": [0, 1] },
                          "created_at": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nueva zona",
        "description": "Crea una nueva zona de entrega",
        "operationId": "createZone",
        "tags": ["Zones"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nombre de la zona (máx 100 caracteres)",
                    "example": "Zona Norte"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Zona creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "zone": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" },
                        "active": { "type": "integer" },
                        "created_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida"
          },
          "409": {
            "description": "Zona con ese nombre ya existe"
          }
        }
      }
    },
    "/zones/{id}": {
      "get": {
        "summary": "Obtener zona específica",
        "description": "Obtiene los detalles de una zona por ID",
        "operationId": "getZoneById",
        "tags": ["Zones"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la zona"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la zona",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "zone": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" },
                        "active": { "type": "integer" },
                        "created_at": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID inválido"
          },
          "404": {
            "description": "Zona no encontrada"
          }
        }
      },
      "patch": {
        "summary": "Actualizar zona",
        "description": "Actualiza el nombre o estado de una zona",
        "operationId": "updateZone",
        "tags": ["Zones"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la zona"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nuevo nombre (opcional)"
                  },
                  "active": {
                    "type": "integer",
                    "enum": [0, 1],
                    "description": "Estado activo (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zona actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "zone": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida"
          },
          "404": {
            "description": "Zona no encontrada"
          }
        }
      }
    },
...
  },
...
}